+++ smol/utils/cluster/container.pyx
@@ -46,13 +46,13 @@ cdef class OrbitContainer:
 
     cpdef public void set_orbits(self, tuple orbit_data) except *:
         """Populated data using a list of orbit data."""
-        cdef int i
+        cdef np.int32_t i
 
         # check that dataypes are correct
         for data in orbit_data:
-            if not isinstance(data[0], int):
+            if not isinstance(data[0], np.int32_t):
                 raise TypeError("id must be an integer.")
-            if not isinstance(data[1], int):
+            if not isinstance(data[1], np.int32_t):
                 raise TypeError("bit_id must be an integer.")
             if not isinstance(data[2], np.ndarray):
                 raise TypeError("correlation_tensors must be a numpy array.")
@@ -83,10 +83,10 @@ cdef class OrbitContainer:
 
     @staticmethod
     cdef OrbitC create_struct(
-            int orbit_id,
-            int bit_id,
-            double[:, ::1] correlation_tensors,
-            long[::1] tensor_indices,
+            np.int32_t orbit_id,
+            np.int32_t bit_id,
+            np.float64_t[:, ::1] correlation_tensors,
+            np.int32_t[::1] tensor_indices,
     ):
         """Set the fields of a OrbitC struct from memoryviews."""
         cdef OrbitC orbit
@@ -131,7 +131,7 @@ cdef class FloatArray2DContainer:
 
     cpdef public void set_arrays(self, tuple arrays) except *:
         """Populated data using a list of 2D arrays."""
-        cdef int i
+        cdef np.int32_t i
 
         for array in arrays:
             if array.ndim != 2:
@@ -154,7 +154,7 @@ cdef class FloatArray2DContainer:
 
 
     @staticmethod
-    cdef FloatArray2D create_struct(double[:, ::1] array):
+    cdef FloatArray2D create_struct(np.float64_t[:, ::1] array):
         """Set the fields of a _FloatArray2D struct from memoryview."""
         cdef FloatArray2D array_struct
         array_struct.size_r = array.shape[0]
@@ -190,7 +190,7 @@ cdef class FloatArray1DContainer:
 
     cpdef public void set_arrays(self, tuple arrays) except *:
         """Populated data using a list of 1D arrays."""
-        cdef int i
+        cdef np.int32_t i
 
         for array in arrays:
             if array.ndim != 1:
@@ -212,7 +212,7 @@ cdef class FloatArray1DContainer:
         self._arrays = arrays
 
     @staticmethod
-    cdef FloatArray1D create_struct(double[::1] array):
+    cdef FloatArray1D create_struct(np.float64_t[::1] array):
         """Set the fields of a FloatArray1D struct from memoryview."""
         cdef FloatArray1D array_struct
         array_struct.size = array.size
@@ -247,7 +247,7 @@ cdef class IntArray1DContainer:
 
     cpdef public void set_arrays(self, tuple arrays) except *:
         """Populated data using a list of 1D arrays."""
-        cdef int i
+        cdef np.int32_t i
 
         for array in arrays:
             if array.ndim != 1:
@@ -269,7 +269,7 @@ cdef class IntArray1DContainer:
         self._arrays = arrays
 
     @staticmethod
-    cdef IntArray1D create_struct(long[::1] array):
+    cdef IntArray1D create_struct(np.int32_t[::1] array):
         """Set the fields of a _FloatArray1D struct from memoryview."""
         cdef IntArray1D array_struct
         array_struct.size = array.size
@@ -304,7 +304,7 @@ cdef class IntArray2DContainer:
 
     cpdef public void set_arrays(self, tuple arrays)  except *:
         """Populated data using a list of 2D arrays."""
-        cdef int i
+        cdef np.int32_t i
 
         for array in arrays:
             if array.ndim != 2:
@@ -326,7 +326,7 @@ cdef class IntArray2DContainer:
         self._arrays = arrays
 
     @staticmethod
-    cdef IntArray2D create_struct(long[:, ::1] array):
+    cdef IntArray2D create_struct(np.int32_t[:, ::1] array):
         """Set the fields of a _IntArray2D struct from memoryview."""
         cdef IntArray2D array_struct
         array_struct.size_r = array.shape[0]
