+++ smol/utils/cluster/evaluator.pxd
@@ -15,62 +15,62 @@ from smol.utils.cluster.container cimport (
 
 @cython.final
 cdef class ClusterSpaceEvaluator(OrbitContainer):
-    cdef int num_corr  # total number of correlation functions
-    cdef int num_orbits  # total number of orbits
-    cdef public double offset  # offset for the interaction tensor
+    cdef np.int32_t num_corr  # total number of correlation functions
+    cdef np.int32_t num_orbits  # total number of orbits
+    cdef public np.float64_t offset  # offset for the interaction tensor
     cdef FloatArray1DContainer cluster_interactions
-    cdef public int num_threads
+    cdef public np.int32_t num_threads
 
     cpdef public void reset_data(
             self,
             tuple orbit_data,
-            int num_orbits,
-            int num_corr_functions,
+            np.int32_t num_orbits,
+            np.int32_t num_corr_functions,
     )
 
     cpdef public void set_cluster_interactions(
-            self, tuple cluster_interaction_tensors, double offset
+            self, tuple cluster_interaction_tensors, np.float64_t offset
     )
 
     cpdef np.ndarray[np.float64_t, ndim=1] correlations_from_occupancy(
             self,
-            const long[::1] occu,
+            const np.int32_t[::1] occu,
             IntArray2DContainer cluster_indices,
     )
 
     cpdef np.ndarray[np.float64_t, ndim=1] interactions_from_occupancy(
             self,
-            const long[::1] occu,
+            const np.int32_t[::1] occu,
             IntArray2DContainer cluster_indices,
     )
 
     cpdef np.ndarray[np.float64_t, ndim=1] delta_correlations_from_occupancies(
             self,
-            const long[::1] occu_f,
-            const long[::1] occu_i,
-            const double[::1] cluster_ratio,
+            const np.int32_t[::1] occu_f,
+            const np.int32_t[::1] occu_i,
+            const np.float64_t[::1] cluster_ratio,
             IntArray2DContainer cluster_indices)
 
     cpdef np.ndarray[np.float64_t, ndim=1] delta_interactions_from_occupancies(
             self,
-            const long[::1] occu_f,
-            const long[::1] occu_i,
-            const double[::1] cluster_ratio,
+            const np.int32_t[::1] occu_f,
+            const np.int32_t[::1] occu_i,
+            const np.float64_t[::1] cluster_ratio,
             IntArray2DContainer cluster_indices
     )
 
     cpdef np.ndarray[np.float64_t, ndim=1] corr_distances_from_occupancies(
             self,
-            const long[::1] occu_f,
-            const long[::1] occu_i,
-            const double[::1] ref_corr_vector,
+            const np.int32_t[::1] occu_f,
+            const np.int32_t[::1] occu_i,
+            const np.float64_t[::1] ref_corr_vector,
             IntArray2DContainer cluster_indices
     )
 
     cpdef np.ndarray[np.float64_t, ndim=1] interaction_distances_from_occupancies(
             self,
-            const long[::1] occu_f,
-            const long[::1] occu_i,
-            const double[::1] ref_interaction_vector,
+            const np.int32_t[::1] occu_f,
+            const np.int32_t[::1] occu_i,
+            const np.float64_t[::1] ref_interaction_vector,
             IntArray2DContainer cluster_indices
     )
